requirements.txt
alembic==1.4.2
amqp==2.5.2
billiard==3.6.3.0
blinker==1.4
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.1.0
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Celery==2.4.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
phonenumbers==8.12.3
protobuf==3.11.3
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
rq==1.3.0
six==1.14.0
SQLAlchemy==1.3.16
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.1
